requirements.txt
openai
streamlit
python-docx
openpyxl